requirements.txt
streamlit
openai-whisper
moviepy
torch
numpy
scipy
ffmpeg-python
python-Levenshtein
pydub
imageio-ffmpeg
